requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
discord==1.7.3
discord-py-slash-command==3.0.1
discord.py==1.7.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==5.0.0
idna==3.3
multidict==5.2.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
